Tag kolla images with %{version}-%{release}
Currently kolla images built with this package have the default tag of
%{upstream_version}, but adding the %{release} will make it clear
which openstack-kolla package generated the images.

I'm not sure how we're going to handle versioning of images when
they're regenerated for new packages, but this change will give the
option of forcing a rebuild by incrementing the Release whenever a
rebuild is required.

Change-Id: I5ebd94b415dc2e6edec275ebb3e1c97eb3d760bb